edgar_svc/requirements.txt
flask-1.1.2
flask-restful-0.3.8
pyedgar-0.1.5
requests-2.25.1
pandas-1.2.2
bs4-0.0.1
flask-cors-3.0.10
geopy-2.1.0